web/requirements.txt
openai
requests
fastapi
uvicorn
python-dotenv
pandas
openpyxl